fix(rpc): tighten wire validation
Reject malformed JSON-RPC envelopes, preserve full UTXO output indexes, reject fractional histogram virtual sizes, and filter blank decoded WebSocket hosts.

Testing: swift test --filter ResponseDecodingValidator/rejectNonJSONRPC2Envelope; swift test --filter ResponseDecodingValidator/rejectEnvelopeWithResultAndError; swift test --filter FulcrumMethodRequestEncodingValidator/encodeRemainingBlockchainRequests; swift test --filter FulcrumMethodRequestEncodingValidator/encodeTypedBlockchainCoreEndpoints; swift test --filter ResponseDecodingValidator/rejectFractionalMempoolFeeHistogramVirtualSize; swift test --filter ClientInterfaceLocalValidator/rejectWhitespaceOnlyWebSocketHost; swift test --filter ServerCatalogRepositoryValidator/rejectWhitespaceOnlyStringCatalogURLHost; swift test --filter ServerCatalogRepositoryValidator/sanitizeFallbackCatalog; swift build

+++ Sources/SwiftFulcrum/PublicAPI/SwiftFulcrum+API.swift
@@ -275,7 +275,7 @@ extension SwiftFulcrum.API.Blockchain {
     public struct UTXO: Sendable {
         public func getInfo(
             transactionHash: String,
-            outputIndex: UInt16
+            outputIndex: UInt
         ) -> SwiftFulcrum.API.Request<SwiftFulcrum.Response.Blockchain.UTXO.GetInfo> {
             .init(method: .blockchain(.utxo(.getInfo(transactionHash: transactionHash, outputIndex: outputIndex))))
         }
